+++ service_catalog/filters/approval_workflow_filter.py
@@ -5,4 +5,4 @@ from service_catalog.models.approval_workflow import ApprovalWorkflow
 class ApprovalWorkflowFilter(SquestFilter):
     class Meta:
         model = ApprovalWorkflow
-        fields = ["name"]
+        fields = ["name", "enabled"]