+++ tests/test_service_catalog/test_api/test_urls/test_approvalworkflow/test_state_machine.py
@@ -8,7 +8,8 @@ class TestServiceCatalogApprovalWorkflowPermissionsEndpoint(BaseTestAPI, TestPer
     def setUp(self):
         super().setUp()
         # Create approval workflow
-        self.approval_worklflow = ApprovalWorkflow.objects.create(operation=self.create_operation_test)
+        self.approval_worklflow = ApprovalWorkflow.objects.create(operation=self.create_operation_test,
+                                                                  enabled=True)
         self.approval_worklflow.scopes.set([self.test_quota_scope])
         self.approval_worklflow.save()